runtime/panic_report: pipe panics & choose exit strategy
This adds a new method `runtimne.SetPanicFd` which sets the UNIX file
descriptor to which the output of a panic will be redirected. This is
useful to persist panics when Go is going to crash, to analyze it during
the next session.

This also adds `runtime.SetExitOnPanic` to let the caller decides how
the runtime should exit when a non recovred panic happens. Setting
`true` will make the runtime call `exit()` and not `raise()` to stop the
process. This is useful to let the caller decides if a Go panic should
be caught by a global signal handler set, for example, by a crash
reporter.