runtime/panic: add runtime.SetGlobalPanicHandler
SetGlobalPanicHandler sets a function which will be called when a
non-recovered panic happens. This lets the caller access debug information
before the crash.

If the caller is trying to set a new global panic handler while another one
is already registered, this function will panic.

If
1. a panic happens while the global panic handler is being called, or
2. a panic happens in another running goroutine while the global panic
handler is being called,

Then, the global panic handler won't be called (again). In other words, the
global panic handler will be called only once in a session.

This function is NOT thread-safe.

IMPORTANT NOTE
This provides a way to perform dynamic operations while the application
should be considered in an inconsistent state: we're about to crash because
of a non-recovered panic. So keep in mind you're walking outside marked paths
and think of the effects of potential inconsistencies in the global program
state.

One would advise to do nothing fancy inthe global panic handler.

USAGE EXAMPLE

    import (
        "fmt",
        "runtime"
    )

    func onPanic(err interface{}) {
    	   fmt.Println("panic reason:", err)
    	   dump := make([]byte, 1*1024*1024) // 1 MB buffer
    	   runtime.Stack(dump, true)
    	   fmt.Println(string(dump))
    }

    func main() {
    	   runtime.SetGlobalPanicHandler(onPanic)

    	   // provoke a panic from a SIGNAL
    	   var i *int
    	   *i = 42

    	   println("program exit normally")
    }

Co-authored-by: Steeve Morin <steeve@zen.ly>
Co-authored-by: Gawen Arab <gawen.arab@c.zen.ly>