cmd/cgo: add JNI's jmethodID/jfieldID to uintptr for Android NDK
In Android's NDK, jmethodID and jfieldID are defined as:

	struct _jfieldID;                       /* opaque structure */
	typedef struct _jfieldID* jfieldID;     /* field IDs */
	struct _jmethodID;                      /* opaque structure */
	typedef struct _jmethodID* jmethodID;   /* method IDs */

Since Go seems them as pointers, a crash can occur at runtime if the GC
tries to dereference those pointers. This is very reproducible in
Android 11 for some reason.

Add those two types to the badJNI types and rework the check logic a
bit.

Fixes #40954

Signed-off-by: Steeve Morin <steeve@zen.ly>